backend/requirements.txt
# Real-Time Sales Assistant - Python Dependencies

# Audio capture
sounddevice>=0.4.6
numpy>=1.24.0

# Speech-to-text
deepgram-sdk>=3.0.0
websockets>=12.0

# LLM
anthropic>=0.34.0

# WebSocket server
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: Local transcription (Phase 4)
# openai-whisper>=20231117

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.5.0
